+++ examples/features/src/timestamp_queries/mod.rs
@@ -181,7 +181,8 @@ impl Queries {
             let timestamp_view = self
                 .destination_buffer
                 .slice(..(size_of::<u64>() as wgpu::BufferAddress * self.num_queries))
-                .get_mapped_range();
+                .get_mapped_range()
+                .unwrap();
             bytemuck::allocation::pod_collect_to_vec(&timestamp_view)
         };