+++ tests/tests/wgpu-gpu/binding_array/buffers.rs
@@ -156,6 +156,7 @@ async fn binding_array_buffers(
         });
         buffer
             .get_mapped_range_mut(..)
+            .unwrap()
             .slice(0..4)
             .copy_from_slice(&data.0);
         buffer.unmap();
@@ -270,7 +271,7 @@ async fn binding_array_buffers(
 
     ctx.device.poll(PollType::wait_indefinitely()).unwrap();
 
-    let data = slice.get_mapped_range();
+    let data = slice.get_mapped_range().unwrap();
 
     assert_eq!(&data[..], &*image);
 }